project_root/backend/requirements.txt
flask
flask_cors
mysql-connector-python
numpy
opencv-python
insightface
vosk
sounddevice
pyttsx3
requests